Normalized isDeleted column, updated migrations and seeds, renamed entity classes, removed duplicate classes, removed AbstractRepository, return type for UuidOrderedTimeGenerator fixed, removed old \Handler references.
