+++ config/autoload/authentication.global.php
@@ -1,14 +1,16 @@
 <?php
 
+use Frontend\User\Entity\User;
+
 return [
     'doctrine' => [
         'authentication' => [
             'orm_default' => [
                 'object_manager' => 'doctrine.entitymanager.orm_default',
-                'identity_class' => \Frontend\User\Entity\User::class,
+                'identity_class' => User::class,
                 'identity_property' => 'identity',
                 'credential_property' => 'password',
-                'credential_callable' => 'Frontend\User\Doctrine\UserAuthentication::verifyCredential',
+                'credential_callable' => 'Frontend\App\Common\UserAuthentication::verifyCredential',
             ],
         ],
     ],
