+++ config/autoload/doctrine.global.php
@@ -3,20 +3,30 @@
 use Ramsey\Uuid\Doctrine\UuidBinaryOrderedTimeType;
 use Ramsey\Uuid\Doctrine\UuidBinaryType;
 use Ramsey\Uuid\Doctrine\UuidType;
+use Frontend\App\Resolver\EntityListenerResolver;
+use Roave\PsrContainerDoctrine\EntityManagerFactory;
+use Doctrine\Common\Persistence\Mapping\Driver\MappingDriverChain;
+use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityManagerInterface;
 
 return [
     'dependencies' => [
         'factories' => [
-            'doctrine.entity_manager.orm_default' => \Roave\PsrContainerDoctrine\EntityManagerFactory::class,
+            'doctrine.entity_manager.orm_default' => EntityManagerFactory::class,
         ],
         'aliases' => [
-            \Doctrine\ORM\EntityManager::class => 'doctrine.entity_manager.orm_default',
-            \Doctrine\ORM\EntityManagerInterface::class => 'doctrine.entity_manager.default',
+            EntityManager::class => 'doctrine.entity_manager.orm_default',
+            EntityManagerInterface::class => 'doctrine.entity_manager.default',
             'doctrine.entitymanager.orm_default' => 'doctrine.entity_manager.orm_default'
         ]
     ],
 
     'doctrine' => [
+        'configuration' => [
+            'orm_default' => [
+                'entity_listener_resolver' => EntityListenerResolver::class,
+            ]
+        ],
         'connection' => [
             'orm_default' => [
                 'doctrine_mapping_types' => [
@@ -29,7 +39,7 @@ return [
             // default metadata driver, aggregates all other drivers into a single one.
             // Override `orm_default` only if you know what you're doing
             'orm_default' => [
-                'class' => \Doctrine\Common\Persistence\Mapping\Driver\MappingDriverChain::class,
+                'class' => MappingDriverChain::class,
                 'drivers' => [],
             ],
         ],
