+++ src/App/src/RoutesDelegator.php
@@ -4,7 +4,6 @@ namespace Frontend\App;
 
 use Fig\Http\Message\RequestMethodInterface;
 use Frontend\App\Controller\LanguageController;
-use Frontend\App\Controller\LanguageHandler;
 use Mezzio\Application;
 use Psr\Container\ContainerInterface;
