--- src/User/src/Doctrine/EntityListenerResolverFactory.php
@@ -1,30 +1,0 @@
-<?php
-
-/**
- * @see https://github.com/dotkernel/frontend/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/frontend/blob/master/LICENSE.md MIT License
- */
-
-declare(strict_types=1);
-
-namespace Frontend\User\Doctrine;
-
-use Frontend\User\Doctrine\EntityListenerResolver;
-use Psr\Container\ContainerInterface;
-
-/**
- * Class EntityListenerResolverFactory
- * @package Frontend\User\Doctrine\Factory
- */
-class EntityListenerResolverFactory
-{
-    /**
-     * @param ContainerInterface $container
-     * @return EntityListenerResolver
-     */
-    public function __invoke(ContainerInterface $container)
-    {
-        return new EntityListenerResolver($container);
-    }
-}
