+++ src/User/src/Repository/UserRoleRepository.php
@@ -4,23 +4,24 @@ declare(strict_types=1);
 
 namespace Frontend\User\Repository;
 
-use Frontend\App\Repository\AbstractRepository;
+use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\NonUniqueResultException;
 use Frontend\User\Entity\UserRole;
 
 /**
  * Class UserRoleRepository
  * @package Frontend\User\Repository
  */
-class UserRoleRepository extends AbstractRepository
+class UserRoleRepository extends EntityRepository
 {
     /**
      * @param string $name
      * @return UserRole|null
-     * @throws \Doctrine\ORM\NonUniqueResultException
+     * @throws NonUniqueResultException
      */
     public function findByName(string $name): ?UserRole
     {
-        $qb = $this->getQueryBuilder();
+        $qb = $this->getEntityManager()->createQueryBuilder();
 
         $qb
             ->select('role')
